dataPipeline/requirements.txt
pandas
psycopg2-binary
country_converter
kagglehub
pycountry